timeline: float draw pipeline kills zoom shimmer; slivers kill strobe
Two zoom-drag artifacts, one root cause each:

1. Sub-pixel clips strobed on/off — integer width rounded to 0 or 1
   depending on grid alignment, which changes with every ppf step.
   Fix: clip_h_span (shared cull+clamp helper, also replaces the
   preview-outline copy) floors draw width at a 1px sliver; boundary
   stripe suppressed on slivers.

2. Boundary lines jiggled ±1px non-monotonically — floor(t*ppf) -
   floor(vs*ppf) steps at different moments per term as ppf sweeps. No
   integer-snapped map can be both pan-rigid and zoom-monotonic, so the
   absolute-grid workaround is deleted: time_to_pixel is now the exact
   float map (t - vs) * ppf, and add_rect/text/line/triangle/waveform
   carry qreal into the already-antialiased painter (QRectF et al).
   Pan rigidity and zoom monotonicity both hold exactly.

Regression tests: test_clip_draw_stability (524-case zoom sweep, 294
failed pre-fix), test_boundary_monotonic_under_zoom (red on the floor
grid); test_clip_width_stable_under_scroll / test_clip_flash_zoomed_out
updated to the float mechanism (epsilon compare, domain rules unchanged).

Authored-By: Joe Shapiro <joe@shapiro.net>
With-Help-From: Claude

+++ src/timeline_renderer.h
@@ -41,13 +41,17 @@ public:
     // Set Lua state for callbacks
     void setLuaState(lua_State* L) { lua_state_ = L; }
 
-    // Drawing command interface (for future Lua integration)
+    // Drawing command interface. Coordinates are qreal: the Lua side maps
+    // time→pixel with an exact linear function (no integer snapping — see
+    // viewport_state.time_to_pixel), and the antialiased painter resolves
+    // fractional coverage. Quantizing here would reintroduce the ±1 px
+    // boundary jiggle during zoom that the float pipeline exists to fix.
     void clearCommands();
-    void addRect(int x, int y, int width, int height, const QString& color);
-    void addText(int x, int y, const QString& text, const QString& color);
-    void addLine(int x1, int y1, int x2, int y2, const QString& color, int width = 1);
-    void addTriangle(int x1, int y1, int x2, int y2, int x3, int y3, const QString& color);
-    void addWaveform(int x, int y, int width, int height,
+    void addRect(qreal x, qreal y, qreal width, qreal height, const QString& color);
+    void addText(qreal x, qreal y, const QString& text, const QString& color);
+    void addLine(qreal x1, qreal y1, qreal x2, qreal y2, const QString& color, int width = 1);
+    void addTriangle(qreal x1, qreal y1, qreal x2, qreal y2, qreal x3, qreal y3, const QString& color);
+    void addWaveform(qreal x, qreal y, qreal width, qreal height,
                      const float* peaks, int peak_count, const QString& color,
                      bool reversed = false);
 
@@ -90,9 +94,9 @@ private:
     // Drawing command structure
     struct DrawCommand {
         enum Type { RECT, TEXT, LINE, TRIANGLE, WAVEFORM } type;
-        int x, y, width, height;
-        int x2, y2; // For lines
-        int x3, y3; // For triangles (third point)
+        qreal x, y, width, height;
+        qreal x2, y2; // For lines
+        qreal x3, y3; // For triangles (third point)
         QString text;
         QColor color;
         int line_width = 1;
